requirements.txt
--trusted-host 4.158.111.250 --extra-index-url https://4.158.111.250/artifactory/api/pypi/pypi-local/simple
core-utils==1.0.0
custom-operators==1.0.0
